requirements.txt
anyio==4.9.0
asgiref==3.8.1
black==25.1.0
certifi==2025.4.26
cfgv==3.4.0
click==8.2.1
colorama==0.4.6
distlib==0.4.0
Django==5.1.2
django-cors-headers==4.5.0
django-extensions==4.1
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
filelock==3.18.0
flake8==7.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
identify==2.6.12
idna==3.10
isort==6.0.1
Markdown==3.7
mccabe==0.7.0
mypy_extensions==1.1.0
mysqlclient==2.2.7
nodeenv==1.9.1
packaging==25.0
pathspec==0.12.1
pillow==12.0.0
platformdirs==4.3.8
pre_commit==4.2.0
psycopg2-binary==2.9.10
pycodestyle==2.14.0
pyflakes==3.4.0
PyJWT==2.9.0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
sniffio==1.3.1
sqlparse==0.5.1
typing_extensions==4.13.2
tzdata==2024.2
virtualenv==20.32.0
